Major refactoring
Allow `initdb` to accept a simple `true` argument which defaults to:

  *  `--auth-local=trust --pgdata=${:pgdata}`

Set the default listen addresses to `''`. I.e. don't listen on TCP by default

Set the default unix socket permissions to 700, allowing connections to
the current user only by default.

Write the default client and server scripts with environment variables
that point to the default host and port, making them simpler and safer
to use.

Add a `bin-directory` option to specify where to write the above scripts,
defaulting to `buildout:bin-directory`.

Run the `cmds` commands with the above mentioned environment
variables, eliminating the need to specify host or port manually.

Make `install` and `upgrade` behave identically.

Add an `uninstall` recipe which stops the database.

Wait for the db to be listening when (re)starting the database inside the
recipe (esp. before running commands) instead of blindly waiting for 4
seconds.

Increase version.

Some Pylint cleanups for good measure.